convert necessary components
