+++ Engine/include/sge.h
@@ -67,4 +67,5 @@
 #include "component/NativeScriptComponent.h"
 #include "component/RenderableComponent.h"
 #include "component/SkyboxComponent.h"
+#include "component/PlayerControllerComponent.h"
 #include "runtime/Prefab.h"
